ci(docs): add fully headless script mode

+++ apps/docs/scripts/generate.sh
@@ -8,21 +8,22 @@ set -euo pipefail              # Strict error handling
 echo "⚙️  Starting generation of static files..."
 echo "💡 You can check .stdout.log and .stderr.log for progress and errors."
 
-processes_to_kill=(node yarn npm)
-echo "⚠️  This script will kill all ${processes_to_kill[*]} processes."
-for process in "${processes_to_kill[@]}"; do
-  pgrep $process | xargs -r kill -9 >/dev/null 2>/dev/null || true
-done
-echo "✅ Killed specified processes."
+if [ "$HELIO_HEADLESS" != "1" ]; then
+  processes_to_kill=(node yarn npm)
+  echo "⚠️  This script will kill all ${processes_to_kill[*]} processes."
+  for process in "${processes_to_kill[@]}"; do
+    pgrep $process | xargs -r kill -9 >/dev/null 2>/dev/null || true
+  done
+  echo "✅ Killed specified processes."
 
-if [ "${BUILD_LATEST:-false}" = "true" ]; then
-  echo "💡 Building latest version in addition to ${PRODUCT_VERSION}"
-else
-  echo "💡 Building only version ${PRODUCT_VERSION}"
+  if [ "${BUILD_LATEST:-false}" = "true" ]; then
+    echo "💡 Building latest version in addition to ${PRODUCT_VERSION}"
+  else
+    echo "💡 Building only version ${PRODUCT_VERSION}"
+  fi
 fi
 
 TARGET_DIR="$(pwd)/.build"     # Define target directory for generated files
-PRINT_DEBUG="${PRINT_DEBUG:-0}"
 
 rm .stdout.log 2>/dev/null || true         # Clear previous logs
 rm .stderr.log 2>/dev/null || true
@@ -32,8 +33,14 @@ mkdir -p "$TARGET_DIR"
 echo "🧹 Emptied target directory: $TARGET_DIR"
 
 echo "💡 Preparing Nuxt for generation..."
-npx nuxi cleanup >> .stdout.log 2>> .stderr.log
-BASE_PATH="/" yarn run nuxt:prepare >> .stdout.log 2>> .stderr.log
+if [ "$HELIO_HEADLESS" -eq 1 ]; then
+  npx nuxi cleanup
+  BASE_PATH="/" yarn run nuxt:prepare
+else
+  npx nuxi cleanup >> .stdout.log 2>> .stderr.log
+  BASE_PATH="/" yarn run nuxt:prepare >> .stdout.log 2>> .stderr.log
+fi
+
 echo "==== end of prepare step ====" >> .stdout.log
 echo "" >> .stdout.log
 echo "✅ Nuxt prepared."
@@ -42,7 +49,7 @@ if [ "$BUILD_LATEST" = "true" ]; then
   echo "💡 Building no-prefix version"
   # 1) Generate for base path "/"
 
-  if [ "$PRINT_DEBUG" -eq 1 ]; then
+  if [ "$HELIO_HEADLESS" -eq 1 ]; then
     BASE_PATH="/" yarn run nuxt:generate
   else
     BASE_PATH="/" yarn run nuxt:generate >> .stdout.log 2>> .stderr.log
@@ -61,7 +68,7 @@ fi
 echo "💡 Building version ${PRODUCT_VERSION}"
 
 # 3) Generate for versioned path "/<PRODUCT_VERSION>/"
-if [ "$PRINT_DEBUG" -eq 1 ]; then
+if [ "$HELIO_HEADLESS" -eq 1 ]; then
   BASE_PATH="/$PRODUCT_VERSION/" yarn run nuxt:generate
 else
   BASE_PATH="/$PRODUCT_VERSION/" yarn run nuxt:generate >> .stdout.log 2>> .stderr.log
